# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Robert AI API documentation.",
        "title": "Robert AI API",
        "termsOfService": "https://robertai.fr/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://robertai.fr/support",
            "email": "support@robertai.fr"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/auth/signup": {
            "post": {
                "description": "Create a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup a new user",
                "parameters": [
                    {
                        "description": "Signup form",
                        "name": "signupForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authHandler.SignupForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/authHandler.SignupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authHandler.SignupError400"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/authHandler.SignupError409"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "authHandler.SignupError400": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Please fully fill in the signup form"
                }
            }
        },
        "authHandler.SignupError409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email not available"
                }
            }
        },
        "authHandler.SignupForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "firstname": {
                    "type": "string",
                    "example": "John"
                },
                "lastname": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "example": "Password123!"
                }
            }
        },
        "authHandler.SignupResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Signup successful"
                }
            }
        },
...
    }
}